ml_model/requirements.txt
fastapi
uvicorn
scikit-learn
transformers
torch
bitsandbytes
peft
pydantic
python-dotenv